add ellipsis for truncated status indicator
