+++ codex-rs/tui/src/bottom_pane/selection_popup_common.rs
@@ -14,6 +14,7 @@ use unicode_width::UnicodeWidthChar;
 use unicode_width::UnicodeWidthStr;
 
 use crate::key_hint::KeyBinding;
+use crate::line_truncation::truncate_line_with_ellipsis_if_overflow;
 use crate::render::Insets;
 use crate::render::RectExt as _;
 use crate::style::user_message_style;
@@ -104,12 +105,6 @@ pub(crate) fn wrap_styled_line<'a>(line: &'a Line<'a>, width: u16) -> Vec<Line<'
     word_wrap_line(line, opts)
 }
 
-fn line_width(line: &Line<'_>) -> usize {
-    line.iter()
-        .map(|span| UnicodeWidthStr::width(span.content.as_ref()))
-        .sum()
-}
-
 fn line_to_owned(line: Line<'_>) -> Line<'static> {
     Line {
         style: line.style,
@@ -122,91 +117,6 @@ fn line_to_owned(line: Line<'_>) -> Line<'static> {
                 content: Cow::Owned(span.content.into_owned()),
             })
             .collect(),
-    }
-}
-
-pub(crate) fn truncate_line_to_width(line: Line<'static>, max_width: usize) -> Line<'static> {
-    if max_width == 0 {
-        return Line::from(Vec::<Span<'static>>::new());
-    }
-
-    let Line {
-        style,
-        alignment,
-        spans,
-    } = line;
-    let mut used = 0usize;
-    let mut spans_out: Vec<Span<'static>> = Vec::new();
-
-    for span in spans {
-        let text = span.content.into_owned();
-        let style = span.style;
-        let span_width = UnicodeWidthStr::width(text.as_str());
-
-        if span_width == 0 {
-            spans_out.push(Span::styled(text, style));
-            continue;
-        }
-
-        if used >= max_width {
-            break;
-        }
-
-        if used + span_width <= max_width {
-            used += span_width;
-            spans_out.push(Span::styled(text, style));
-            continue;
-        }
-
-        let mut truncated = String::new();
-        for ch in text.chars() {
-            let ch_width = UnicodeWidthChar::width(ch).unwrap_or(0);
-            if used + ch_width > max_width {
-                break;
-            }
-            truncated.push(ch);
-            used += ch_width;
-        }
-
-        if !truncated.is_empty() {
-            spans_out.push(Span::styled(truncated, style));
-        }
-
-        break;
-    }
-
-    Line {
-        style,
-        alignment,
-        spans: spans_out,
-    }
-}
-
-pub(crate) fn truncate_line_with_ellipsis_if_overflow(
-    line: Line<'static>,
-    max_width: usize,
-) -> Line<'static> {
-    if max_width == 0 {
-        return Line::from(Vec::<Span<'static>>::new());
-    }
-
-    let width = line_width(&line);
-    if width <= max_width {
-        return line;
-    }
-
-    let truncated = truncate_line_to_width(line, max_width.saturating_sub(1));
-    let Line {
-        style,
-        alignment,
-        mut spans,
-    } = truncated;
-    let ellipsis_style = spans.last().map(|span| span.style).unwrap_or_default();
-    spans.push(Span::styled("…", ellipsis_style));
-    Line {
-        style,
-        alignment,
-        spans,
     }
 }
