+++ codex-rs/tui/src/lib.rs
@@ -80,6 +80,7 @@ mod get_git_diff;
 mod history_cell;
 pub mod insert_history;
 mod key_hint;
+mod line_truncation;
 pub mod live_wrap;
 mod markdown;
 mod markdown_render;
